+++ Repo/user/quake/r_misc.c
@@ -442,8 +442,8 @@ r_refdef.viewangles[2]=    0;
 	{
 		if (r_dowarp)
 		{
-			if ((vid.width <= vid.maxwarpwidth) &&
-				(vid.height <= vid.maxwarpheight))
+			if (((int)vid.width <= vid.maxwarpwidth) &&
+				((int)vid.height <= vid.maxwarpheight))
 			{
 				vrect.x = 0;
 				vrect.y = 0;
@@ -520,4 +520,3 @@ r_refdef.viewangles[2]=    0;
 
 	D_SetupFrame ();
 }
-
